Fixed renaming for story archive folders
